requirements.txt
asgiref==3.8.1
Django==4.2.9
django-admin-volt==1.0.10
django-cors-headers==4.4.0
django-debug-toolbar==4.4.2
django-filter==24.2
django-grappelli==4.0.1
django-jazzmin==3.0.0
django-restframework==0.0.1
django-tinymce==4.1.0
django-unfold==0.5.0
djangorestframework==3.15.2
gunicorn==21.2.0
packaging==24.1
pillow==10.3.0
python-dotenv==1.0.1
sqlparse==0.5.0
str2bool==1.1
tzdata==2024.1
whitenoise==6.6.0
